% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotnum.R
\name{barplotnum}
\alias{barplotnum}
\title{Barplot of each cluster for numerical variables data set.}
\usage{
barplotnum(dataori, clust, nc = 1)
}
\arguments{
\item{dataori}{An original data set.}

\item{clust}{A vector of cluster membership.}

\item{nc}{A number of column of the plot.}
}
\value{
Function returns a barplot.
}
\description{
This function create a barplot from a cluster result.
}
\details{
This is a function to produce a barplot for each cluster.
}
\examples{
dat <- iris[,1:4]
memb <- cutree(hclust(dist(dat)),3)
barplotnum(dat, memb)
barplotnum(dat, memb, 2)

}
\author{
Weksi Budiaji \cr Contact: \email{budiaji@untirta.ac.id}
}
