% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clust4}
\alias{clust4}
\title{4 clustered data}
\format{A data frame with 300 rows and 3 variables:
\describe{
  \item{x1}{X1.}
  \item{x2}{X2.}
  \item{class}{Class membership.}
}}
\usage{
clust4
}
\description{
A dataset containing two variables of 300 objects and their class memberships
generated by a clusterGeneration package.
}
\keyword{datasets}
