% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastkmed.R
\name{fastkmed}
\alias{fastkmed}
\title{A simple and fast k-medoid algorithm from Park and Jun.}
\usage{
fastkmed(distdata, ncluster, iterate = 10)
}
\arguments{
\item{distdata}{A matrix of distance objects (n x n) or a diss class.}

\item{ncluster}{A number of cluster.}

\item{iterate}{A number of iteration for clustering algorithm.}
}
\value{
Function returns a partitioning clustering algorithm result consists of cluster
membership, cluster medoid, the minimum distance to the cluster medoid.
}
\description{
This function computes and returns the clustering result
computed by using a specified distance via Park and Jun's k-medoids.
}
\details{
This is a two-data-set equivalent of the standard function of distance.
It returns a matrix of all pairwise distances between rows in data1 and data2. \cr
For a list of available numerical distance, binary and categorical distance, one can
search by \code{distname()}
}
\examples{
num <- as.matrix(iris[,1:4])
mrwdist <- distNumeric(num, num, method = "mrw")
result <- fastkmed(mrwdist, ncluster = 3, iterate = 50)
table(result$cluster, iris[,5])


}
\references{
Park, H., Jun, C., 2009. A simple and fast algorithm for k-medoids clustering.
Expert Systems with Applications 36, 3336-3341.
}
\author{
Weksi Budiaji \cr Contact: \email{budiaji@untirta.ac.id}
}
