% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{globalfood}
\alias{globalfood}
\title{Global food security index}
\format{A data frame with 113 rows and 4 variables:
\describe{
  \item{affordability}{affordability}
  \item{availability}{availability}
  \item{safety}{safety}
  \item{resilience}{resilience}
}}
\source{
\url{http://foodsecurityindex.eiu.com}
}
\usage{
globalfood
}
\description{
A dataset containing four variables of 113 countries for their food security
index based on panelists evaluation.
}
\keyword{datasets}
