% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{matching}
\alias{matching}
\title{A pair distance for binary/ categorical variables.}
\usage{
matching(x, y)
}
\arguments{
\item{x}{A data frame/ matrix.}

\item{y}{A second data frame/ matrix.}
}
\description{
This function computes and returns the distance matrix computed by using
the simple matrching distance.
}
\details{
This is a function to compute simple matching distance.
It returns a matrix of distance objects, i.e n x n.
}
\examples{
set.seed(1)
a <- matrix(sample(1:2, 7*3, replace = TRUE), 7, 3)
matching(a, a)

}
\author{
Weksi Budiaji
}
