% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow.R
\name{shadow}
\alias{shadow}
\title{Centroid shadow value (CSV) index of each cluster based on medoid}
\usage{
shadow(distdata, idmedoid, idcluster)
}
\arguments{
\item{distdata}{A distance object/ a n x n distance matrix.}

\item{idmedoid}{A vector of id medoids.}

\item{idcluster}{A vector of cluster membership.}
}
\description{
This function is deprecated, use the \code{\link{csv}}
function instead.
}
