% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhoutte.R
\name{silhoutte}
\alias{silhoutte}
\title{Silhoutte index of each cluster}
\usage{
silhoutte(distdata, idmedoid, idcluster)
}
\arguments{
\item{distdata}{A distance object/ a n x n distance matrix.}

\item{idmedoid}{A vector of id medoids.}

\item{idcluster}{A vector of cluster membership.}
}
\description{
This function is deprecated, use the \code{\link{sil}}
function instead.
}
