% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepkmed.R
\name{stepkmed}
\alias{stepkmed}
\title{Step k-medoid algorithm from Yu et al.}
\usage{
stepkmed(distdata, ncluster, iterate = 10, alpha = 1)
}
\arguments{
\item{distdata}{A matrix of distance objects (n x n) or a diss class.}

\item{ncluster}{A number of clusters.}

\item{iterate}{A number of iterations for the clustering algorithm.}

\item{alpha}{A numeric number to determine the range of initial medoids
selection.}
}
\description{
This function is deprecated, use the \code{\link{inckmed}}
function instead.
}
